Arc camera system (#24)
* Working Arc Ball Camera Rotations

This commit adds the zooming, and rotating needed for arc ball. There is no panning yet available and it is not based on the position in the transform component for the camera yet.

* Working Arc Ball Camera Rotations

This commit adds the zooming, and rotating needed for arc ball. There is no panning yet available and it is not based on the position in the transform component for the camera yet.

* Zoom Properties, Flipped the world right side up, Fixed Oval Rotations

The zoom properties have been added to the editor so that you can control how far out or in you want to zoom before you are stopped. The world was upside down for some reason, so that is now fixed, and I fixed the oval Arcball camera rotations.

* Added Panning controls for the camera

Also fixed some final errors and everything is now working and ready to go.