Fixed audio bug, general code cleaning.
I've removed the enum_cast macro with a template that can (optionally, by default it uses std::size_t) take a type to cast the enum value to.

Also removed the YTEUnusedArgument macro and replaced all uses with the UnusedArguments function, consolidating uses where possible.

The main menu music wasn't playing due to the PlayEvent call being commented out by Nick or I, uncommented it and fixed.

Configured the assimp dependency in cmake to remove some of it's default additions.