Fix facial animation bug after merging the Mesh/VkMesh separation
The facial animation was only updating the Mesh, not the graphics side mesh. This should get fleshed out further, as right now we're going through the hacky InstantiatedModel interface. This interface isn't ideal because it's meant to represent an instance. Perhaps the Mesh could call a function on the Renderer to let it know it's been updated?