Fixed a water bug
In the first commit of this series I implemented a relatively simple optimization to the initial water UBOs::Model generation. This accidentally missed testing by me, and had a major bug in it. Thankfully, the optimization was still valid, just misunderstood/misread a piece of the translation code. Easy fix once discovered.